requirements_cuda.txt
onnxruntime-gpu
# numpy
sounddevice
levenshtein
loguru
jinja2
requests
textual
pyyaml
librosa
pillow
keyboard
pyautogui
mouse